Moved GameSpecificFixups.js into the subdir with the other related files
